# docs/openapi.yaml
basePath: /api/v1
definitions:
  apierror.Response:
    properties:
      code:
        type: string
        description: Stable machine-readable error code. Clients should branch on this field rather than
          message text.
      details:
        additionalProperties:
          type: string
        type: object
        description: Optional field-level validation details keyed by request field.
      message:
        type: string
        description: Safe human-readable error summary.
      request_id:
        type: string
        description: Server-generated ID for tracing this request; also returned in X-Request-ID.
    type: object
    description: Response representation used by this API.
    example:
      code: INVALID_REQUEST
      message: Invalid request
      request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      details:
        total: must be a decimal string
...
  event.SpinResponse:
    properties:
      candidates:
        items:
          $ref: '#/definitions/model.CandidatePreviewDto'
        type: array
      reward:
        type: string
        description: Exact-decimal amount represented as a JSON string with two fractional digits in responses.
      slots:
        items:
          type: string
        type: array
      stealToken:
        $ref: '#/definitions/model.StealTokenDto'
    type: object
...
  model.CandidatePreviewDto:
    properties:
      group_id:
        type: string
        description: Identifier of the associated resource.
      index:
        type: integer
      name:
        type: string
      role_id:
        type: string
        description: Identifier of the associated resource.
    type: object
...
  model.DailyRewardScheduleItem:
    properties:
      amount:
        type: string
        description: Exact-decimal amount represented as a JSON string with two fractional digits in responses.
      date:
        type: string
        description: Reward date in DD-MM-YYYY format.
    type: object
  model.DailyRewardScheduleResponse:
    properties:
      default_amount:
        type: string
        description: Exact-decimal amount represented as a JSON string with two fractional digits in responses.
      overrides:
        items:
          $ref: '#/definitions/model.DailyRewardScheduleItem'
        type: array
    type: object
...
  model.SetDailyRewardRequest:
    properties:
      amount:
        type: string
        description: Nonnegative reward as a decimal JSON string; numeric JSON tokens are rejected.
    required:
    - amount
    type: object
...
  model.StealTokenDto:
    properties:
      expires_at:
        type: string
        format: date-time
        description: Timestamp in RFC 3339 format.
      message:
        type: string
      token:
        type: string
      victim_count:
        type: integer
    type: object
...
host: localhost:8080
info:
  title: Intania 888 API
  version: 1.0.0-breaking
  description: "Browser and application API for Intania 888. All application routes use the /api/v1 base\
    \ path. Browser authentication uses an HttpOnly session cookie; call GET /auth/me after login to retrieve\
    \ the current profile and session-bound csrf_token. Send that token in X-CSRF-Token with browser mutations\
    \ while a session is active and send an Origin allowed by the server. Mutations are rejected when\
    \ either check fails. \n\nMoney values are JSON strings. Responses use exactly two fractional digits;\
    \ requests accept up to two. Rate values are JSON strings with up to six fractional digits in requests\
    \ and six in responses. Numeric JSON tokens are not accepted for these decimal values. \n\nAPI failures\
    \ use the shared apierror.Response shape: code, message, request_id, and optional field-level details.\
    \ The same request ID is returned in the X-Request-ID header. Dates and timestamps use the formats\
    \ documented by each field. The separate /external/* backend integration is described in docs/README.md\
    \ and is intentionally outside this browser-focused reference.\n\nThe API is rate limited to 200 requests\
    \ per minute per client IP. Exceeding the limit returns 429 TOO_MANY_REQUESTS."
paths:
...
  /events/daily-rewards:
    get:
      description: Administrator-only. Returns the configured default daily reward and the date-specific
        overrides currently in effect.
      produces:
      - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/model.DailyRewardScheduleResponse'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '401':
          description: 'unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '403':
          description: 'admin access required Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: FORBIDDEN
              message: The request is not allowed
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '500':
          description: 'Failed to list daily reward schedule Failure body follows apierror.Response; code
            is stable, message is safe to display, and request_id matches the X-Request-ID response header.
            Failure body follows apierror.Response: stable code, safe message, request_id, and optional
            field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INTERNAL_ERROR
              message: Internal server error
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: List daily reward schedule
      tags:
      - Event
  /events/daily-rewards/{date}:
    delete:
      description: Administrator-only. Removes the override for the supplied DD-MM-YYYY date so the configured
        default reward applies again.
      parameters:
      - description: Reward date in DD-MM-YYYY format.
        in: path
        name: date
        required: true
        type: string
        pattern: ^\d{2}-\d{2}-\d{4}$
      - description: Session-bound token returned by GET /auth/me. Required for browser mutations.
        in: header
        name: X-CSRF-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        '200':
          description: Daily reward override deleted
          schema:
            additionalProperties:
              type: string
            type: object
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '400':
          description: 'Invalid reward date Failure body follows apierror.Response; code is stable, message
            is safe to display, and request_id matches the X-Request-ID response header. Failure body
            follows apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INVALID_REQUEST
              message: Invalid request
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '401':
          description: 'unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '403':
          description: 'admin access required Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: FORBIDDEN
              message: The request is not allowed
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '404':
          description: 'Daily reward override not found Failure body follows apierror.Response; code is
            stable, message is safe to display, and request_id matches the X-Request-ID response header.
            Failure body follows apierror.Response: stable code, safe message, request_id, and optional
            field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
        '500':
          description: 'Failed to delete daily reward override Failure body follows apierror.Response;
            code is stable, message is safe to display, and request_id matches the X-Request-ID response
            header. Failure body follows apierror.Response: stable code, safe message, request_id, and
            optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INTERNAL_ERROR
              message: Internal server error
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Delete a daily reward override
      tags:
      - Event
    put:
      consumes:
      - application/json
      description: Administrator-only. Creates or replaces the reward override for a DD-MM-YYYY date.
        Repeating the same request is safe.
      parameters:
      - description: Reward date in DD-MM-YYYY format.
        in: path
        name: date
        required: true
        type: string
        pattern: ^\d{2}-\d{2}-\d{4}$
      - description: Daily reward amount
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SetDailyRewardRequest'
      - description: Session-bound token returned by GET /auth/me. Required for browser mutations.
        in: header
        name: X-CSRF-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        '200':
          description: Set daily reward successful
          schema:
            additionalProperties:
              type: string
            type: object
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '400':
          description: 'Invalid request payload Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INVALID_REQUEST
              message: Invalid request
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '401':
          description: 'unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '403':
          description: 'admin access required Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: FORBIDDEN
              message: The request is not allowed
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '500':
          description: 'Failed to set daily reward Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INTERNAL_ERROR
              message: Internal server error
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Set daily reward
      tags:
      - Event
  /events/redeem/daily:
    get:
      consumes:
      - application/json
      description: Claims the configured daily reward for the authenticated account. Each account can
        claim once per reward day; a repeated claim returns a conflict.
      produces:
      - application/json
      responses:
        '200':
          description: redeemed daily reward successful
          schema:
            additionalProperties:
              type: string
            type: object
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '401':
          description: 'unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '404':
          description: 'user not found Failure body follows apierror.Response; code is stable, message
            is safe to display, and request_id matches the X-Request-ID response header. Failure body
            follows apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
        '409':
          description: 'daily reward already claimed Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
        '500':
          description: 'internal server error Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INTERNAL_ERROR
              message: Internal server error
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Redeem daily reward
      tags:
      - Event
  /events/spin/slot:
    post:
      description: Spends one supported stake amount (50.00, 100.00, or 500.00) on a slot-machine spin.
        The server checks the account balance and returns the committed reward, slot result, and any token
        granted.
      parameters:
      - description: Stake amount as a decimal string. Accepted values are 50.00, 100.00, and 500.00.
        in: query
        name: spendAmount
        required: true
        type: string
      - description: Session-bound token returned by GET /auth/me. Required for browser mutations.
        in: header
        name: X-CSRF-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        '200':
          description: slot result
          schema:
            $ref: '#/definitions/event.SpinResponse'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '400':
          description: 'invalid spend amount or user profile Failure body follows apierror.Response; code
            is stable, message is safe to display, and request_id matches the X-Request-ID response header.
            Failure body follows apierror.Response: stable code, safe message, request_id, and optional
            field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INVALID_REQUEST
              message: Invalid request
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '401':
          description: 'unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '404':
          description: 'user not found Failure body follows apierror.Response; code is stable, message
            is safe to display, and request_id matches the X-Request-ID response header. Failure body
            follows apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
        '422':
          description: 'insufficient balance Failure body follows apierror.Response; code is stable, message
            is safe to display, and request_id matches the X-Request-ID response header. Failure body
            follows apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
        '500':
          description: 'internal server error Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INTERNAL_ERROR
              message: Internal server error
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '403':
          description: 'The caller lacks the required permission, or an Origin or CSRF check failed. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: FORBIDDEN
              message: The request is not allowed
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Spin the slot machine
      tags:
      - Event
...
schemes:
- http
securityDefinitions:
  CookieSession:
    description: Browser session cookie. The cookie is HttpOnly and is normally sent by the browser automatically.
      Production uses the __Host-session cookie name. Swagger 2.0 represents this as a Cookie header.
      For browser mutations, also send X-CSRF-Token from GET /auth/me and an allowed Origin.
    in: header
    name: Cookie
    type: apiKey
swagger: '2.0'
consumes:
- application/json
produces:
- application/json
...